requirements.txt
httpx[http2]
pycryptodome
